// src/app/core/api/model/chatMessageDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ChatMessageDto { 
    role: ChatMessageDto.RoleEnum;
    content: string;
}
export namespace ChatMessageDto {
    export const RoleEnum = {
        System: 'system',
        User: 'user',
        Assistant: 'assistant'
    } as const;
    export type RoleEnum = typeof RoleEnum[keyof typeof RoleEnum];
}